Refactor: -rename variable index to currentIndex in Game.cs -move timer.Elapsed definitions to the constructor of Game.cs -remove Clamp method when calculating durationMinutes in CalculateWPM method in Game.cs -change the type of variables GrossWPM and NetWPM to float in Result.cs -cast the values for GrossWPM and NetWPM to int when displaying results in GUI.cs -method DisplayResults now expects a Result object as its parameter -rename method ClearConsole to Clear in GUI.cs -rename method StartGame to SetupGameScreen in GUI.cs

+++ Result.cs
@@ -3,8 +3,8 @@
     public class Result
     {
         public int CharactersTyped { get; set; }
-        public int GrossWPM { get; set; }
-        public int NetWPM { get; set; }
+        public float GrossWPM { get; set; }
+        public float NetWPM { get; set; }
         public List<char> Mistakes { get; set; }
 
         public Result()
